+++ web/modules/custom/qs_badge/tests/src/Kernel/BadgeManagerCountEventsByDatesTest.php
@@ -20,7 +20,7 @@ final class BadgeManagerCountEventsByDatesTest extends ResoliKernelTestBase {
   /**
    * {@inheritdoc}
    */
-  public static $modules = [
+  protected static $modules = [
     'qs_badge',
   ];